renamed ipa_canopen_master to canopen_master
