+++ canopen_master/include/canopen_master/exceptions.h
