+++ canopen_master/include/canopen_master/master.h
@@ -1,7 +1,7 @@
 #ifndef H_IPA_CANOPEN_MASTER
 #define H_IPA_CANOPEN_MASTER
 
-#include <ipa_canopen_master/canopen.h>
+#include <canopen_master/canopen.h>
 
 #include <boost/interprocess/sync/interprocess_condition.hpp>
 #include <boost/interprocess/sync/interprocess_mutex.hpp>
@@ -323,7 +323,7 @@ class SharedMaster: public Master{
     boost::shared_ptr<ipa_can::CommInterface> interface_;
 public:
     SharedMaster(const std::string &name, boost::shared_ptr<ipa_can::CommInterface> interface)
-    : name_("ipa_canopen_master_shm_"+name), remover_(name_.c_str()),
+    : name_("canopen_master_shm_"+name), remover_(name_.c_str()),
         managed_shm_(boost::interprocess::open_or_create, name_.c_str(), 4096),
         interface_(interface)  {}
     virtual boost::shared_ptr<SyncLayer> getSync(const SyncProperties &properties);
