+++ canopen_master/src/objdict.cpp
@@ -1,4 +1,4 @@
-#include <ipa_canopen_master/objdict.h>
+#include <canopen_master/objdict.h>
 #include <boost/property_tree/ptree.hpp>
 #include <boost/property_tree/ini_parser.hpp>
 #include <boost/foreach.hpp>
