+++ ipa_canopen_chain_ros/include/ipa_canopen_chain_ros/chain_ros.h
@@ -1,9 +1,9 @@
 #ifndef H_IPA_CANOPEN_CHAIN_ROS
 #define H_IPA_CANOPEN_CHAIN_ROS
 
-#include <ipa_canopen_master/canopen.h>
-#include <ipa_canopen_master/master.h>
-#include <ipa_canopen_master/can_layer.h>
+#include <canopen_master/canopen.h>
+#include <canopen_master/master.h>
+#include <canopen_master/can_layer.h>
 #include <ros/ros.h>
 #include <ros/package.h>
 #include <cob_srvs/Trigger.h>
