+++ ipa_canopen_test/src/test.cpp
@@ -1,5 +1,5 @@
-#include <ipa_canopen_master/canopen.h>
-#include <ipa_canopen_master/master.h>
+#include <canopen_master/canopen.h>
+#include <canopen_master/master.h>
 #include <boost/make_shared.hpp>
 #include <iostream>
